requirements.txt
langchain
langchain-openai
python-dotenv
sqlalchemy
psycopg2-binary 